Fix mag_reports JSON build warning on 64-bit Linux
ts_sent_ms is an int64_t, which is long long on macOS but long on
64-bit Linux. Printing it with %lld is correct on the Mac dev host --
clean under Apple clang and gcc alike -- but warns under gcc on the
ground machine, where int64_t is long. Cast the argument to long long
so the format matches on both.

+++ utils/mag_reports.c
@@ -910,7 +910,7 @@ static void print_json(const rec_t *r)
     if (r->command[0]) printf("\"command\":\"%s\",", cmd_esc);
     else               printf("\"command\":null,");
 
-    if (r->ts_sent_ms) printf("\"ts_sent_ms\":%lld,", r->ts_sent_ms);
+    if (r->ts_sent_ms) printf("\"ts_sent_ms\":%lld,", (long long)r->ts_sent_ms);
     else               printf("\"ts_sent_ms\":null,");
 
     printf("\"meas_time\":\"%s\",\"meas_basis\":\"%s\",", meas_iso, r->meas_basis);
